+++ src/utilities/isPlainObject.ts
@@ -1,0 +1,3 @@
+export function isPlainObject(value: any): value is Object {
+  return Object.prototype.toString.call(value) === "[object Object]";
+}
