+++ src/utilities/useMounted.ts
@@ -1,0 +1,13 @@
+import * as React from "react";
+
+export function useMounted(): boolean {
+  const ref = React.useRef<boolean>(true);
+
+  React.useEffect(() => {
+    return () => {
+      ref.current = false;
+    };
+  }, []);
+
+  return ref.current;
+}
